REFACTOR: moved all passenger and three3gs to single app
