--- three_gs/admin.py
@@ -1,5 +1,0 @@
-from django.contrib import admin
-from .models import ThreeGs
-
-# Register your models here.
-admin.site.register(ThreeGs)
