--- three_gs/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class ThreeGsConfig(AppConfig):
-    name = 'three_gs'
